Use a std::set for replication queue dedup.
This has a couple of O(log n) ops (where n == queue len), but is safer
than the O(1) lock-free option that has almost no memory overhead I
was considering.

That may still be possible, but it will require a lot of lock analysis
to determine whether the new algorithm would be lock free because it's
safe or because the right things weren't locked at the right time.

So in the meantime, we'll go with the slightly slower, but safer and
more isolated approach.